Move tests into separate files
* Easier to navigate when spec files are smaller
* Move files that are always required to `test_helper`
