+++ test/joiner_join_snippet_data_test.rb
@@ -1,8 +1,5 @@
 require_relative 'test_helper'
-require_relative '../lib/team_api'
-require_relative 'site'
 
-require 'minitest/autorun'
 require 'weekly_snippets/version'
 
 module TeamApi
