+++ test/joiner_team_index_by_field_test.rb
@@ -1,9 +1,4 @@
 require_relative 'test_helper'
-require_relative '../lib/team_api'
-require_relative 'site'
-
-require 'jekyll/document'
-require 'minitest/autorun'
 
 module TeamApi
   # Tests via impl.team_by_email for convenience, as it's a thin wrapper
