+++ test/test_helper.rb
@@ -1,3 +1,8 @@
+require_relative '../lib/team_api'
+require_relative 'site'
+
+require 'minitest/autorun'
+
 if ENV['TRAVIS']
   require 'codeclimate-test-reporter'
   CodeClimate::TestReporter.start
